All functional test fixed after core refactoring
