--- config/testing/original-image-cache.yml
@@ -1,43 +1,0 @@
-imports:
-    - { resource: 'services/core.yml' }
-    - { resource: 'services/imaging.yml' }
-    - { resource: 'services/utilities.yml' }
-    - { resource: 'services/filesystem-storage.yml' }
-
-parameters:
-    filesystem_image_cache.base_directory: '%test.web_directory%'
-    filesystem_storage_driver.base_directory: '%test.filesource_directory%'
-
-services:
-    logger:
-        class: Psr\Log\NullLogger
-    security:
-        class: Strider2038\ImgCache\Core\NullSecurity
-
-    # imaging
-    filesystem_image_storage:
-        class: Strider2038\ImgCache\Imaging\ImageStorage
-        arguments:
-            - '@original_image_extractor'
-            - '@original_image_writer'
-
-    # imaging/extraction
-    original_image_extractor:
-        class: Strider2038\ImgCache\Imaging\Extraction\SourceImageExtractor
-        arguments:
-            - '@plain_filename_parser'
-            - '@filesystem_storage_accessor'
-            - '@image_processor'
-
-    # imaging/insertion
-    original_image_writer:
-        class: Strider2038\ImgCache\Imaging\Insertion\SourceImageWriter
-        arguments:
-            - '@plain_filename_parser'
-            - '@filesystem_storage_accessor'
-
-    # imaging/parsing
-    plain_filename_parser:
-        class: Strider2038\ImgCache\Imaging\Parsing\Filename\PlainFilenameParser
-        arguments:
-            - '@entity_validator'
