+++ config/testing/services/geo-map-parameters-parser.yml
@@ -1,4 +1,4 @@
 imports:
-    - { resource: 'services/geo-map.yml' }
+    - { resource: 'services/geo-map-storage.yml' }
     - { resource: 'services/imaging.yml' }
     - { resource: 'services/utilities.yml' }
