+++ config/testing/services/imagick-image-processor.yml
