--- config/testing/test.yml
@@ -1,23 +1,0 @@
-imports:
-  - { resource: 'services/core.yml' }
-  - { resource: 'services/configuration.yml' }
-  - { resource: 'services/utilities.yml' }
-  - { resource: 'services/imaging.yml' }
-  - { resource: 'services/request-handling.yml' }
-  - { resource: 'services/image-cache.yml' }
-  - { resource: 'services/parameters.yml' }
-  - { resource: 'local.yml', ignore_errors: true }
-
-parameters:
-    server_configuration:
-        - ~
-    configuration_filename: ''
-
-services:
-    configuration_loader:
-        class: Strider2038\ImgCache\Configuration\ConfigurationLoader
-        arguments:
-            - '@configuration_file_parser'
-            - '@configuration_processor'
-            - '@configuration_factory'
-            - '%configuration_filename%'
